replaced older wallclock_diff with new timer

+++ pol-core/pol/uimport.h
@@ -9,7 +9,7 @@ Notes
 */
 
 int read_data();
-int write_data( unsigned int& dirty_writes, unsigned int& clean_writes, unsigned int& elapsed_ms );
+int write_data( unsigned int& dirty_writes, unsigned int& clean_writes, long long& elapsed_ms );
 void read_account_data();
 void read_starting_locations();
 void read_gameservers();
